[TextEditor] support margins after TextViewMargin
Width and Offset calculations break when a new
margin is added after the TextViewMargin.

This is due to the fact that the first margin
with a negative Width[Request] would be given
the rest of the allocation width.

As such, margins would not be rendered if they
were inserted after the TextViewMargin.

This patch provides support for margins that
follow variable width margins. Multiple variable
width margins are also supported by subdividing
variable allocated space evenly across any
variable margin instances.

It might be nice in the future to build on this
support by allowing margins to specify their
desired percentage widths, but this is likely
to be a very uncommon situation, so the simple
even subdivision approach is taken now.

Additionally, Margin.Width is renamed to
Margin.WidthRequest (abstract), and Margin.Width
is re-introduced (not abstract) to be set by
the TextArea after computing widths based on
Margin.WidthRequest.

Rendering by the margins themselves remains
unaffected (Draw methods should continue to
use Margin.Width), so all logic changes are
contained in TextArea.